gracefully handle deleting modelMax through input
When modelMax is deleted, e.g. by a user pushing backspace when focused on an input field which has ng-model set to modelMin or modelMax, then modelMax is are set to 0 or modelMin.

This fixes the issue that happens when trying to delete the last cipher from an input field - the value just resets to max.